+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `Slytherin` will be documented in this file.
 
 ### Fixed
 - Type hinting of all classes using `PHPStan`
+- Checking of available instances in `Container::value`
 
 ## [0.9.6](https://github.com/rougin/slytherin/compare/v0.9.5...v0.9.6) - 2023-11-16